requirements.txt
streamlit>=1.28.0
openai>=1.1.0
pinecone-client>=2.2.0
requests>=2.28.0